Refactoring and Javadocs
