+++ community/cypher/cypher-compiler-2.3/src/main/scala/org/neo4j/cypher/internal/compiler/v2_3/planner/SemanticTable.scala
@@ -53,6 +53,10 @@ class SemanticTable(
       throw new InternalException(s"Did not find any type information for identifier $s", e)
   }
 
+  def isNode(expr: String) = getTypeFor(expr) == symbols.CTNode.invariant
+
+  def isRelationship(expr: String) = getTypeFor(expr) == symbols.CTRelationship.invariant
+
   def isNode(expr: Identifier) = types(expr).specified == symbols.CTNode.invariant
 
   def isRelationship(expr: Identifier) = types(expr).specified == symbols.CTRelationship.invariant
